+++ R/read-psi-spectrapen-csv.r
@@ -29,7 +29,7 @@
 #' @references \url{https://psi.cz/}
 #'
 #' @examples
-#'
+#'  # fetch path to example file to read
 #'  file.name <-
 #'    system.file("extdata", "spectrum-psi-spectrapen-SP.csv",
 #'                package = "photobiologyInOut", mustWork = TRUE)