+++ NEWS.md
@@ -13,8 +13,9 @@ of data files are welcome for these and other instruments).
 
 -   Replace use of deprecated `tidyr::gather_()`.
 -   Add function `read_cid_spectravue_csv()` supporting data import from measurements CSV files output by CID Bio-Science's SpectraVue CI-710s leaf spectrometer.
-_   Add function `read_foreign2mspct()` to apply an import function to a
+-   Add function `read_foreign2mspct()` to apply an import function to a
 list of files.
+-   Add color-related functions `spct_CRI()`, `spct_CCT()` and `expect_SSI()` implemented as wrappers to functions in package 'colorSpec' that accept `source_spct` objects as arguments. CRI, CCT and SSI are quantities used to describe light used for illumination, photography, cinematography and video.
 
 # photobiologyInOut 0.4.23